renamed container directory to features
