+++ webapp/app/(auth)/confirm/page.tsx
@@ -1,4 +1,4 @@
-import { ConfirmSignupForm } from "@/components/containers/ConfirmSignupForm/ConfirmSignupForm"
+import { ConfirmSignupForm } from "@/components/features/ConfirmSignupForm/ConfirmSignupForm"
 
 // メールアドレス認証コード入力画面
 export default function ConfirmPage() {
